Adds options for performing backup on a remote host through SSH.
User can specify remote hostname using '-r' option and username (if it is not
	equal to current username) using '-u' option.

Actually backup of multiple directories are handled by issuing multiple rsync
commands. So, if a password is required for SSH login, user is prompted for it
once for each directory. I will try to find a workaround for this issue.